+++ src/dockerfiles/Automattic__wp-calypso-21492.Dockerfile
@@ -52,6 +52,7 @@ ENV DBUS_SESSION_BUS_ADDRESS="unix:path=/run/dbus/system_bus_socket"
 RUN dbus-daemon --system --fork
 
 ENV PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true
+ENV PUPPETEER_SKIP_DOWNLOAD=true
 ENV OPENSSL_CONF /etc/ssl
 
 RUN useradd -m chromeuser
@@ -91,33 +92,32 @@ python2 -V
 EOF_f4fd4a31eaa9
 
 
-RUN <<EOF_96328218397b
+RUN <<EOF_ae345dd1fa0c
 #!/bin/bash
 set -euxo pipefail
 git clone -o origin https://github.com/Automattic/wp-calypso /testbed
-chmod -R 777 /testbed
 cd /testbed
 git reset --hard 97d16987f9b4897e3e71ab9d3688ad59d4c57e3d
 git remote remove origin
-TARGET_EPOCH=$(git show -s --format=%ct 97d16987f9b4897e3e71ab9d3688ad59d4c57e3d)
-git tag -l | while read tag; do TAG_COMMIT=$(git rev-list -n 1 "$tag"); TAG_EPOCH=$(git show -s --format=%ct "$TAG_COMMIT"); if [ "$TAG_EPOCH" -gt "$TARGET_EPOCH" ]; then git tag -d "$tag"; fi; done
-git branch -D $(git branch | grep -v "^\*") 2>/dev/null || true
+git branch | grep -v '^\*' | xargs -r git branch -D || true
+git tag -l | while read tag; do   git merge-base --is-ancestor "$tag" HEAD 2>/dev/null || git tag -d "$tag" >/dev/null; done
 git reflog expire --expire=now --all
+git gc --prune=now --aggressive
+TARGET_EPOCH=$(git show -s --format=%ct 97d16987f9b4897e3e71ab9d3688ad59d4c57e3d)
+AFTER_EPOCH=$((TARGET_EPOCH + 1))
+AFTER_TIMESTAMP=$(date -u -d "@$AFTER_EPOCH" "+%Y-%m-%d %H:%M:%S")
+COMMIT_COUNT=$(git log --oneline --all --since="$AFTER_TIMESTAMP" | wc -l)
+[ "$COMMIT_COUNT" -eq 0 ] || exit 1
 cd - || true
+chmod -R 777 /testbed
 cd /testbed
 git clean -fdxq
 source $NVM_DIR/nvm.sh
 npm install --unsafe-perm
-EOF_96328218397b
+npm install cpf@1.0.1 hoek@6.1.3 --no-save --no-prune --legacy-peer-deps || true
+EOF_ae345dd1fa0c
 
 
-RUN <<EOF_da37027e7ea5
-#!/bin/bash
-set -euxo pipefail
-mkdir -p /swebench/image_assets
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/A0Nsu9q2Uw.png' 'https://cldup.com/A0Nsu9q2Uw.png' || true
-EOF_da37027e7ea5
-
+COPY src/image_assets/Automattic__wp-calypso-21492/ /swebench/image_assets/
 
 WORKDIR /testbed
