+++ src/dockerfiles/Automattic__wp-calypso-21977.Dockerfile
@@ -52,6 +52,7 @@ ENV DBUS_SESSION_BUS_ADDRESS="unix:path=/run/dbus/system_bus_socket"
 RUN dbus-daemon --system --fork
 
 ENV PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true
+ENV PUPPETEER_SKIP_DOWNLOAD=true
 ENV OPENSSL_CONF /etc/ssl
 
 RUN useradd -m chromeuser
@@ -91,33 +92,32 @@ python2 -V
 EOF_f4fd4a31eaa9
 
 
-RUN <<EOF_41d9a2a7a4ff
+RUN <<EOF_a2378e731fe4
 #!/bin/bash
 set -euxo pipefail
 git clone -o origin https://github.com/Automattic/wp-calypso /testbed
-chmod -R 777 /testbed
 cd /testbed
 git reset --hard 6c46667b4d171af667fa6fc0990c0cf35e954ae3
 git remote remove origin
-TARGET_EPOCH=$(git show -s --format=%ct 6c46667b4d171af667fa6fc0990c0cf35e954ae3)
-git tag -l | while read tag; do TAG_COMMIT=$(git rev-list -n 1 "$tag"); TAG_EPOCH=$(git show -s --format=%ct "$TAG_COMMIT"); if [ "$TAG_EPOCH" -gt "$TARGET_EPOCH" ]; then git tag -d "$tag"; fi; done
-git branch -D $(git branch | grep -v "^\*") 2>/dev/null || true
+git branch | grep -v '^\*' | xargs -r git branch -D || true
+git tag -l | while read tag; do   git merge-base --is-ancestor "$tag" HEAD 2>/dev/null || git tag -d "$tag" >/dev/null; done
 git reflog expire --expire=now --all
+git gc --prune=now --aggressive
+TARGET_EPOCH=$(git show -s --format=%ct 6c46667b4d171af667fa6fc0990c0cf35e954ae3)
+AFTER_EPOCH=$((TARGET_EPOCH + 1))
+AFTER_TIMESTAMP=$(date -u -d "@$AFTER_EPOCH" "+%Y-%m-%d %H:%M:%S")
+COMMIT_COUNT=$(git log --oneline --all --since="$AFTER_TIMESTAMP" | wc -l)
+[ "$COMMIT_COUNT" -eq 0 ] || exit 1
 cd - || true
+chmod -R 777 /testbed
 cd /testbed
 git clean -fdxq
 source $NVM_DIR/nvm.sh
 npm install --unsafe-perm
-EOF_41d9a2a7a4ff
+npm install cpf@1.0.1 hoek@6.1.3 --no-save --no-prune --legacy-peer-deps || true
+EOF_a2378e731fe4
 
 
-RUN <<EOF_c69d20f0699b
-#!/bin/bash
-set -euxo pipefail
-mkdir -p /swebench/image_assets
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/32190568-6dde3198-bd6b-11e7-96cd-81522b663880.png' 'https://user-images.githubusercontent.com/1595739/32190568-6dde3198-bd6b-11e7-96cd-81522b663880.png' || true
-EOF_c69d20f0699b
-
+COPY src/image_assets/Automattic__wp-calypso-21977/ /swebench/image_assets/
 
 WORKDIR /testbed
