+++ src/dockerfiles/Automattic__wp-calypso-22782.Dockerfile
@@ -52,6 +52,7 @@ ENV DBUS_SESSION_BUS_ADDRESS="unix:path=/run/dbus/system_bus_socket"
 RUN dbus-daemon --system --fork
 
 ENV PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true
+ENV PUPPETEER_SKIP_DOWNLOAD=true
 ENV OPENSSL_CONF /etc/ssl
 
 RUN useradd -m chromeuser
@@ -91,33 +92,31 @@ python2 -V
 EOF_9e43395224ce
 
 
-RUN <<EOF_8f357bb9ce9f
+RUN <<EOF_3e4b9565bfa4
 #!/bin/bash
 set -euxo pipefail
 git clone -o origin https://github.com/Automattic/wp-calypso /testbed
-chmod -R 777 /testbed
 cd /testbed
 git reset --hard 5f4c22c561a31972c973dc9dadd907836070bceb
 git remote remove origin
-TARGET_EPOCH=$(git show -s --format=%ct 5f4c22c561a31972c973dc9dadd907836070bceb)
-git tag -l | while read tag; do TAG_COMMIT=$(git rev-list -n 1 "$tag"); TAG_EPOCH=$(git show -s --format=%ct "$TAG_COMMIT"); if [ "$TAG_EPOCH" -gt "$TARGET_EPOCH" ]; then git tag -d "$tag"; fi; done
-git branch -D $(git branch | grep -v "^\*") 2>/dev/null || true
+git branch | grep -v '^\*' | xargs -r git branch -D || true
+git tag -l | while read tag; do   git merge-base --is-ancestor "$tag" HEAD 2>/dev/null || git tag -d "$tag" >/dev/null; done
 git reflog expire --expire=now --all
+git gc --prune=now --aggressive
+TARGET_EPOCH=$(git show -s --format=%ct 5f4c22c561a31972c973dc9dadd907836070bceb)
+AFTER_EPOCH=$((TARGET_EPOCH + 1))
+AFTER_TIMESTAMP=$(date -u -d "@$AFTER_EPOCH" "+%Y-%m-%d %H:%M:%S")
+COMMIT_COUNT=$(git log --oneline --all --since="$AFTER_TIMESTAMP" | wc -l)
+[ "$COMMIT_COUNT" -eq 0 ] || exit 1
 cd - || true
+chmod -R 777 /testbed
 cd /testbed
 git clean -fdxq
 source $NVM_DIR/nvm.sh
 npm install --unsafe-perm
-EOF_8f357bb9ce9f
+EOF_3e4b9565bfa4
 
 
-RUN <<EOF_96b7fed1e55f
-#!/bin/bash
-set -euxo pipefail
-mkdir -p /swebench/image_assets
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/35518737-9d088bb6-04cf-11e8-9334-74b20dd2b6d7.png' 'https://user-images.githubusercontent.com/66797/35518737-9d088bb6-04cf-11e8-9334-74b20dd2b6d7.png' || true
-EOF_96b7fed1e55f
-
+COPY src/image_assets/Automattic__wp-calypso-22782/ /swebench/image_assets/
 
 WORKDIR /testbed
